backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
pydantic
python-dotenv
beautifulsoup4
requests
langfuse
